refactor: create package `serializers`
