+++ jit-compiler/src/main/java/org/aya/compiler/serializers/CallKind.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.compiler;
+package org.aya.compiler.serializers;
 
 import org.aya.compiler.free.FreeUtil;
 import org.aya.syntax.core.def.ConDefLike;
