+++ jit-compiler/src/main/java/org/aya/compiler/serializers/ExprializeUtil.java
@@ -1,13 +1,13 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.compiler;
+package org.aya.compiler.serializers;
 
 import com.intellij.openapi.util.text.StringUtil;
 import kala.collection.SeqLike;
 import kala.collection.immutable.ImmutableSeq;
 import org.jetbrains.annotations.NotNull;
 
-public interface ExprializeUtils {
+public interface ExprializeUtil {
   String SEP = ", ";
 
   static @NotNull String makeNew(@NotNull String className, String... terms) {
