+++ jit-compiler/src/main/java/org/aya/compiler/serializers/ModuleSerializer.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.compiler;
+package org.aya.compiler.serializers;
 
 import kala.collection.immutable.ImmutableSeq;
 import org.aya.compiler.free.FreeClassBuilder;
@@ -12,7 +12,7 @@ import org.jetbrains.annotations.NotNull;
 
 import java.lang.constant.ClassDesc;
 
-import static org.aya.compiler.NameSerializer.getReference;
+import static org.aya.compiler.serializers.NameSerializer.getReference;
 
 /**
  * Serializing a module, note that it may not a file module, so we need not to make importing.
