+++ jit-compiler/src/main/java/org/aya/compiler/serializers/PatternSerializer.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2024 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.compiler;
+package org.aya.compiler.serializers;
 
 import kala.collection.SeqView;
 import kala.collection.immutable.ImmutableSeq;
