+++ jit-compiler/src/test/java/RedBlackTreeTest.java
@@ -22,7 +22,7 @@ import java.util.Random;
 import java.util.function.Function;
 import java.util.function.IntFunction;
 
-import static org.aya.compiler.NameSerializer.getClassName;
+import static org.aya.compiler.serializers.NameSerializer.getClassName;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class RedBlackTreeTest {
